# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GDial Emergency Notification System API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/burn-messages": {
      "post": {
        "tags": [
          "burn-messages"
        ],
        "summary": "Create Burn Message",
        "description": "Create a new burn message that can be viewed once and then deleted.",
        "operationId": "create_burn_message_api_burn_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BurnMessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BurnMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/burn/{token}": {
      "get": {
        "tags": [
          "burn-messages"
        ],
        "summary": "View Burn Message",
        "description": "View a burn message. This will mark the message as viewed and it cannot be accessed again.",
        "operationId": "view_burn_message_burn__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/burn-sms": {
      "post": {
        "tags": [
          "burn-messages"
        ],
        "summary": "Send Burn Sms",
        "description": "Send an SMS with a link to a burn message.",
        "operationId": "send_burn_sms_api_burn_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BurnSmsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/burn-message-call": {
      "post": {
        "tags": [
          "burn-messages"
        ],
        "summary": "Initiate Burn Message Call",
        "description": "Initiate voice calls that direct recipients to a burn message.",
        "operationId": "initiate_burn_message_call_api_burn_message_call_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BurnVoiceCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ringbot/audio/{file_name}": {
      "get": {
        "summary": "Get Audio File",
        "description": "Serve a generated audio file for TTS playback.",
        "operationId": "get_audio_file_ringbot_audio__file_name__get",
        "parameters": [
          {
            "name": "file_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BurnMessageCreate": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "expires_in_hours": {
            "type": "integer",
            "title": "Expires In Hours",
            "default": 24
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "BurnMessageCreate",
        "description": "Schema for creating a burn message"
      },
      "BurnMessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "viewed": {
            "type": "boolean",
            "title": "Viewed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "token",
          "content",
          "created_at",
          "expires_at",
          "viewed"
        ],
        "title": "BurnMessageResponse",
        "description": "Schema for returning a burn message"
      },
      "BurnSmsRequest": {
        "properties": {
          "recipients": {
            "items": {
              "type": "string",
              "format": "uuid4"
            },
            "type": "array",
            "title": "Recipients",
            "default": []
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "message_content": {
            "type": "string",
            "title": "Message Content"
          },
          "custom_link_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Link Text"
          },
          "expires_in_hours": {
            "type": "integer",
            "title": "Expires In Hours",
            "default": 24
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          }
        },
        "type": "object",
        "required": [
          "message_content"
        ],
        "title": "BurnSmsRequest",
        "description": "Schema for sending an SMS with a burn message link"
      },
      "BurnVoiceCallRequest": {
        "properties": {
          "recipients": {
            "items": {
              "type": "string",
              "format": "uuid4"
            },
            "type": "array",
            "title": "Recipients",
            "default": []
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "burn_content": {
            "type": "string",
            "title": "Burn Content"
          },
          "intro_message": {
            "type": "string",
            "title": "Intro Message"
          },
          "custom_link_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Link Text"
          },
          "dtmf_digit": {
            "type": "string",
            "title": "Dtmf Digit",
            "default": "1"
          },
          "dtmf_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dtmf Message"
          },
          "expires_in_hours": {
            "type": "integer",
            "title": "Expires In Hours",
            "default": 24
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          }
        },
        "type": "object",
        "required": [
          "burn_content",
          "intro_message"
        ],
        "title": "BurnVoiceCallRequest",
        "description": "Schema for initiating a voice call that directs to a burn message"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}